+++ np.MeetingNotes/src/NPMeetingNotes.js
@@ -4,6 +4,7 @@ const scriptLoad = new Date()
 
 import moment from 'moment-business-days'
 
+import { getTemplate } from '../../np.Templating/lib/core'
 import pluginJson from '../plugin.json'
 import { showMessage, chooseFolder, chooseOption } from '../../helpers/userInput'
 import { getNoteByFilename } from '../../helpers/note'
@@ -11,8 +12,6 @@ import { isCalendarNoteFilename } from '@helpers/regex'
 import { log, logDebug, logError, clo, JSP, timer } from '@helpers/dev'
 import { findProjectNoteUrlInText } from '@helpers/urls'
 import { getAttributes } from '@helpers/NPFrontMatter'
-import NPTemplating from 'NPTemplating'
-
 /**
  * Insert a template into a daily note or the current editor.
  * Called from the 'insert template' button of an empty note.
@@ -31,22 +30,19 @@ export async function insertNoteTemplate(origFileName: string, dailyNoteDate: Da
   let templateContent = DataStore.projectNoteByFilename(templateFilename)?.content
 
   if (!templateContent) {
-    logError(pluginJson, `couldnt load content of template "${templateFilename}", try NPTemplating method`)
-    templateContent = await NPTemplating.getTemplate(templateFilename)
+    logError(pluginJson, `couldnt load content of template "${templateFilename}"`)
+    templateContent = await getTemplate(templateFilename)
     //
     // templateContent = await DataStore.invokePluginCommandByName('getTemplate', 'np.Templating', [templateFilename])
     return
   }
 
-  logDebug(pluginJson, 'preRender() template')
-  const { frontmatterBody, frontmatterAttributes } = await NPTemplating.preRender(templateContent)
+  logDebug(pluginJson, 'calling renderFrontmatter() to pre-render template')
+  const { frontmatterBody, frontmatterAttributes } = await DataStore.invokePluginCommandByName('renderFrontmatter', 'np.Templating', [templateContent])
 
-  // const { frontmatterBody, frontmatterAttributes } = await DataStore.invokePluginCommandByName('preRender', 'np.Templating', [templateContent])
+  logDebug(pluginJson, `render() template with frontmatterAttributes: [${Object.keys(frontmatterAttributes).join(', ')}]`)
 
-  logDebug(pluginJson, 'render() template')
-  const result = await NPTemplating.render(frontmatterBody, frontmatterAttributes)
-
-  // const result = await DataStore.invokePluginCommandByName('render', 'np.Templating', [frontmatterBody, frontmatterAttributes])
+  const result = await DataStore.invokePluginCommandByName('render', 'np.Templating', [frontmatterBody, frontmatterAttributes])
 
   if (dailyNoteDate) {
     logDebug(pluginJson, `apply rendered template to daily note with date ${String(dailyNoteDate)}`)
@@ -67,18 +63,6 @@ export async function insertNoteTemplate(origFileName: string, dailyNoteDate: Da
       Editor.insertTextAtCursor(result)
     }
   }
-}
-
-/**
- * Insert a template into a daily note or the current editor.
- * Called from the 'insert template' button of an empty note.
- * Called from 'insert template' context menu or "/-commands"
- * @param {string} origFileName -> (optional) Template filename, if not set the user will be asked
- * @param {Date} dailyNoteDate -> (optional) Date of the daily note, if not set the current editor will be used
- */
-export async function renderNoteTemplate(origFileName: string, dailyNoteDate: Date): Promise<string> {
-  logDebug(pluginJson, 'renderNoteTemplate')
-  return 'hello world'
 }
 
 /**
@@ -108,6 +92,7 @@ async function askWhichNoteToOpen(filenames: Array<string>): Promise<string> {
  * @param {string} template
  */
 export async function newMeetingNoteFromID(eventID: string, template?: string): Promise<void> {
+  const startTime = new Date()
   try {
     logDebug(pluginJson, `${timer(scriptLoad)} - newMeetingNoteFromID id:${eventID} template:${String(template)}`)
     const selectedEvent: TCalendarItem = await Calendar.eventByID(eventID)
@@ -153,6 +138,7 @@ export async function newMeetingNoteFromID(eventID: string, template?: string): 
   } catch (error) {
     logError(pluginJson, `error in newMeetingNoteFromID: ${JSP(error)}`)
   }
+  logDebug(pluginJson, `newMeetingNoteFromID: finished after ${timer(startTime)}`)
 }
 
 /**
@@ -177,17 +163,27 @@ async function selectEventAndTemplate(
  * @returns {Promise<{result: string, attrs: any}>}
  */
 async function renderTemplateForEvent(selectedEvent, templateFilename): Object {
-  let templateData, templateContent
+  logDebug(pluginJson, `${timer(scriptLoad)} - renderTemplateForEvent: templateFilename: "${templateFilename}"; selectedEvent.title: "${selectedEvent?.title}"`)
+  let templateVariables, templateContent
   if (selectedEvent) {
-    templateData = generateTemplateData(selectedEvent)
+    templateVariables = generateEventData(selectedEvent)
   }
   if (templateFilename) {
     templateContent = DataStore.projectNoteByFilename(templateFilename)?.content || ''
   }
-  const { frontmatterBody, frontmatterAttributes } = await NPTemplating.preRender(templateContent, templateData)
+  logDebug(
+    pluginJson,
+    `${timer(scriptLoad)} - renderTemplateForEvent: calling renderFrontmatter() with content and variables (${
+      templateVariables ? Object.keys(templateVariables).join(', ') : 'none'
+    })`,
+  )
+  const { frontmatterBody, frontmatterAttributes } = await DataStore.invokePluginCommandByName('renderFrontmatter', 'np.Templating', [templateContent, templateVariables])
+
   clo(frontmatterBody, 'renderTemplateForEvent frontmatterBody:')
   clo(frontmatterAttributes, 'renderTemplateForEvent frontmatterAttributes:')
-  const result = await NPTemplating.render(frontmatterBody, frontmatterAttributes)
+
+  const result = await DataStore.invokePluginCommandByName('render', 'np.Templating', [frontmatterBody, frontmatterAttributes, templateVariables])
+
   return { result, attrs: frontmatterAttributes }
 }
 
@@ -316,7 +312,7 @@ async function createNoteAndLinkEvent(selectedEvent: TCalendarItem | null, rende
  */
 export async function newMeetingNote(_selectedEvent?: TCalendarItem, _templateFilename?: string, forceNewNote: boolean = false): Promise<void> {
   const { selectedEvent, templateFilename } = await selectEventAndTemplate(_selectedEvent, _templateFilename)
-  logDebug(pluginJson, `${timer(scriptLoad)} - newMeetingNote: got selectedEvent and templateFilename`)
+  logDebug(pluginJson, `${timer(scriptLoad)} - newMeetingNote: got selectedEvent and templateFilename; calling renderTemplateForEvent()`)
   const { result, attrs } = await renderTemplateForEvent(selectedEvent, templateFilename)
   logDebug(pluginJson, `${timer(scriptLoad)} - newMeetingNote: rendered template`)
   clo(result, 'rendered template:')
@@ -576,17 +572,24 @@ const errorReporter = async (error: any, note: TNote) => {
  */
 async function chooseTemplateIfNeededFromTemplateTitle(templateTitle?: string, onlyMeetingNotes: boolean = false): Promise<?string> {
   // Get the filename and then pass to the main function
-  logDebug(pluginJson, `${timer(scriptLoad)} - chooseTemplateIfNeededFromTemplateTitle starting`)
+  logDebug(
+    pluginJson,
+    `${timer(scriptLoad)} - chooseTemplateIfNeededFromTemplateTitle starting with templateTitle: "${String(templateTitle)}" and onlyMeetingNotes: ${String(onlyMeetingNotes)}`,
+  )
   if (templateTitle) {
-    const matchingTemplates = DataStore.projectNotes.filter((n) => n.title === templateTitle)
-    logDebug(pluginJson, `${timer(scriptLoad)}- got ${matchingTemplates.length} template matches from '${templateTitle}'`)
-    if (matchingTemplates && matchingTemplates.length > 0) {
-      return await chooseTemplateIfNeeded(matchingTemplates[0].filename, onlyMeetingNotes)
+    if (!templateTitle.endsWith('.md') && !templateTitle.endsWith('.txt')) {
+      const matchingTemplates = DataStore.projectNotes.filter((n) => n.title === templateTitle)
+      logDebug(pluginJson, `${timer(scriptLoad)}- got ${matchingTemplates.length} template title matches for '${templateTitle}'`)
+      if (matchingTemplates && matchingTemplates.length > 0) {
+        logDebug(pluginJson, `${timer(scriptLoad)}- choosing the first template title match for '${templateTitle}'; filename: ${matchingTemplates[0].filename}`)
+        return await chooseTemplateIfNeeded(matchingTemplates[0].filename, onlyMeetingNotes)
+      }
     } else {
-      return await chooseTemplateIfNeeded(templateTitle, onlyMeetingNotes)
+      logDebug(pluginJson, `${timer(scriptLoad)}- we have a filename (template name ends with .md or .txt), so sending filename directly to chooseTemplateIfNeeded`)
     }
+    return await chooseTemplateIfNeeded(templateTitle, onlyMeetingNotes)
   }
-  logDebug(pluginJson, `${timer(scriptLoad)} - chooseTemplateIfNeededFromTemplateTitle ending`)
+  logDebug(pluginJson, `${timer(scriptLoad)} - no template title provided; chooseTemplateIfNeededFromTemplateTitle ending`)
   return await chooseTemplateIfNeeded('', onlyMeetingNotes)
 }
 
@@ -599,7 +602,10 @@ async function chooseTemplateIfNeededFromTemplateTitle(templateTitle?: string, o
  * @returns {Promise<string>} filename of the template
  */
 async function chooseTemplateIfNeeded(templateFilename?: string, onlyMeetingNotes: boolean = false): Promise<?string> {
-  logDebug(pluginJson, `${timer(scriptLoad)} - chooseTemplateIfNeeded`)
+  logDebug(
+    pluginJson,
+    `${timer(scriptLoad)} - chooseTemplateIfNeeded() starting with templateFilename: "${String(templateFilename)}" and onlyMeetingNotes: ${String(onlyMeetingNotes)}`,
+  )
   try {
     if (!templateFilename) {
       logDebug(pluginJson, `${timer(scriptLoad)} - no template was defined, find all available templates and show them`)
@@ -651,7 +657,7 @@ async function chooseTemplateIfNeeded(templateFilename?: string, onlyMeetingNote
           : { index: 0 }
       return templates[selectedTemplate.index].filename
     } else {
-      logDebug(pluginJson, `will use Template file '${templateFilename}' ...`)
+      logDebug(pluginJson, `Will use Template file '${templateFilename}'`)
     }
     return templateFilename
   } catch (error) {
@@ -702,16 +708,16 @@ async function chooseEventIfNeeded(selectedEvent?: TCalendarItem | null): Promis
 }
 
 /**
- * Creates template data as input for np.Templating to parse a template.
+ * Creates event data object (properties/methods) to be used as input for np.Templating to parse a template.
  * @param {TCalendarItem} selectedEvent
  * @returns {Object} data and methods for the template
  */
-function generateTemplateData(selectedEvent: TCalendarItem): { data: Object, methods: Object } {
+function generateEventData(selectedEvent: TCalendarItem): { data: Object, methods: Object } {
   if (!selectedEvent) {
-    logError(pluginJson, 'generateTemplateData: no event provided')
+    logError(pluginJson, 'generateEventData: no event provided')
     return { data: {}, methods: {} }
   }
-  logDebug(pluginJson, `generateTemplateData running for event titled: "${selectedEvent.title}"`)
+  logDebug(pluginJson, `generateEventData populating event details for event titled: "${selectedEvent?.title}"`)
   return {
     data: {
       eventTitle: selectedEvent.title,
@@ -722,8 +728,11 @@ function generateTemplateData(selectedEvent: TCalendarItem): { data: Object, met
       eventAttendeeNames: selectedEvent && selectedEvent?.attendees?.length ? selectedEvent.attendeeNames.join(', ') : '',
       eventLocation: selectedEvent.location, // not yet documented!
       eventCalendar: selectedEvent.calendar,
+      eventDateValue: selectedEvent.date,
+      eventEndDateValue: selectedEvent.endDate,
     },
     methods: {
+      // NOTE: functions cannot be passed via DataStore.invokePluginCommandByName(), so we will have to create these methods in np.Templating render()
       eventDate: (format: string = 'YYYY MM DD') => {
         return moment(selectedEvent.date).format(`${format}`)
       },